Add files via upload
The provided code serves the purpose of uploading onto a D1mini, which is a microcontroller based on the Esp8266 platform. Its primary function is to automate the operation of a garage door when the D1mini enters a designated range of a certain Wi-Fi network. IFTTT and Blynk.console is integrated to this project.
This code can be accessed using Arduino IDE.